// world-cup-api/src/match/model/match-formation.enum.ts
export enum MatchFormation {
  F_4_3_3 = '4-3-3',
  F_4_4_2 = '4-4-2',
  F_4_2_3_1 = '4-2-3-1',
  F_3_5_2 = '3-5-2',
  F_3_4_3 = '3-4-3',
  F_4_1_4_1 = '4-1-4-1',
  F_4_5_1 = '4-5-1',
  F_5_3_2 = '5-3-2',
  F_5_4_1 = '5-4-1',
  F_4_1_2_1_2 = '4-1-2-1-2',
  F_4_3_2_1 = '4-3-2-1',
  F_3_4_1_2 = '3-4-1-2',
  F_4_4_1_1 = '4-4-1-1',
  F_4_3_1_2 = '4-3-1-2',
  F_3_1_4_2 = '3-1-4-2',
}